Add finetune notebook for LLAMA

+++ pretrained_llama/src/config.py
@@ -10,7 +10,7 @@ LLAMA32_CONFIG = {
     "hidden_dim": 8192,  # Size of the intermediate dimension in FeedForward
     "n_kv_groups": 8,  # Key-Value groups for grouped-query attention
     "rope_base": 500_000.0,  # The base in RoPE's "theta"
-    "dtype": torch.bfloat16,  # Lower-precision dtype to reduce memory usage
+    "dtype": torch.float8_e5m2,  # Lower-precision dtype to reduce memory usage
     "rope_freq": {  # RoPE frequency scaling
         "factor": 32.0,
         "low_freq_factor": 1.0,
